rt_linux.c: Multiple rtnl_lock related changes and more...
Removed RtmpOSNetDevProtect, it is no longer needed, being
replaced by calls to rtnl_lock, rtnl_unlock.

RtmpOSNetDevAttach: Modified. Replaced the code

	if (rtnl_locked) {
		rtnl_lock();
		ret = register_netdevice(pNetDev);
		rtnl_unlock();
	} else {
		ret = register_netdev(pNetDev);
	}

Considering the code is identical to (check how register_netdev is implemented):

        if (rtnl_locked) {
                ret = register_netdev(pNetDev);
        } else {
                ret = register_netdev(pNetDev);
        }

Then the whole sequence is equivalent to:

	ret = register_netdev(pNetDev);

This also makes the rtnl_locked variable useless and by extension the
structure field pDevOpHook->needProtcted unneeded, as the field was only used in
this routine. I believe the original purpose was to avoid locking an already
locked state. So being paranoid, I added an ASSERT if we try to lock an
already locked state.

Also modified RtmpOSNetDevDetact, replacing the sequence
	...
	rtnl_lock();
	unregister_netdevice(pNetDev);
	rtnl_unlock();
	...

by an equivalent
	...
	nregister_netdev(pNetDev);
	...

And as usual, plenty of additional code cleanup.

Signed-off-by: Juro Bystricky <jurobystricky@hotmail.com>